+++ mmtf-serialization/src/main/java/org/rcsb/mmtf/serialization/quickmessagepackdeserialization/BC.java
@@ -1,4 +1,4 @@
-package org.rcsb.mmtf.serialization.mp;
+package org.rcsb.mmtf.serialization.quickmessagepackdeserialization;
 
 /**
  * byte converter
