+++ mmtf-serialization/src/main/java/org/rcsb/mmtf/serialization/quickmessagepackdeserialization/BinaryDocument.java
@@ -21,7 +21,7 @@
  *  License along with this library; if not, write to the Free Software
  *  Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-package org.rcsb.mmtf.serialization.mp;
+package org.rcsb.mmtf.serialization.quickmessagepackdeserialization;
 
 
 import java.io.BufferedInputStream;
