+++ mmtf-serialization/src/main/java/org/rcsb/mmtf/serialization/quickmessagepackdeserialization/GenericOutputChannel.java
@@ -1,4 +1,4 @@
-package org.rcsb.mmtf.serialization.mp;
+package org.rcsb.mmtf.serialization.quickmessagepackdeserialization;
 
 public interface GenericOutputChannel {
